+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Use new pypi.org links [#295](https://github.com/etalab/udata-gouvfr/pull/295)
 - Ensure active users have a confirmed_at date (migration) [#298](https://github.com/etalab/udata-gouvfr/pull/298)
 - Remove credits page [#306](https://github.com/etalab/udata-gouvfr/pull/306)
+- Fix `modal-lg` width [#311](https://github.com/etalab/udata-gouvfr/pull/311)
 
 ## 1.3.2 (2018-03-28)