+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 - Fix 404/missing css on front pages [#1709](https://github.com/opendatateam/udata/pull/1709)
 - Fix markdown max image width (front) [#1707](https://github.com/opendatateam/udata/pull/1707)
 - Adds ETag to internal avatar for efficient caching control [#1712](https://github.com/opendatateam/udata/pull/1712)
+- Show resource type in modal (front) [#1714](https://github.com/opendatateam/udata/pull/1714)
 
 ## 1.3.12 (2018-05-31)